change block height from uint64 to int64 & lint (#21)
* change block height from uint64 to int64 & lint

* update readme

* bug checksum length checker

+++ executor/host/handler.go
@@ -20,7 +20,7 @@ func (h *Host) beginBlockHandler(_ context.Context, args nodetypes.BeginBlockArg
 
 func (h *Host) endBlockHandler(_ context.Context, args nodetypes.EndBlockArgs) error {
 	// collect more msgs if block height is not latest
-	blockHeight := uint64(args.Block.Header.Height)
+	blockHeight := args.Block.Header.Height
 	msgQueue := h.GetMsgQueue()
 	if blockHeight != args.LatestHeight && len(msgQueue) > 0 && len(msgQueue) <= 10 {
 		return nil
